Organize LineChart types and exports
