+++ projects/js-packages/charts/src/components/line-chart/index.tsx
@@ -1,8 +1,9 @@
 export { default as LineChart } from './line-chart';
 export type {
+	AnnotationStyles,
+	LineChartAnnotationProps,
 	RenderLineStartGlyphProps,
 	LineChartProps,
 	TooltipDatum,
 	CurveType,
-} from './line-chart';
-export type { AnnotationStyles } from './types';
+} from './types';
